Move base out of aoc2022
I moved the base directory out of aoc2022 so that it can be reused for
future years. One issue is that I have been sort of iterating on the
base code so if I continue doing this it could break previous puzzle
solutions that rely on a particular version of the base code.

Maybe it is better to leave the base in aoc2022 and then next year I can
copy over the base and any modifications only affect that year's
directory of puzzles.

I am undecided though so moving for now.